pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.5",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "nltk>=3.9.1",
    "numpy>=2.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pypdf2>=3.0.1",
    "python-barcode>=0.15.1",
    "qrcode>=8.2",
    "bcrypt>=4.3.0",
    "requests>=2.32.4",
    "pymupdf>=1.26.3",
    "reportlab>=4.4.2",
    "werkzeug>=3.1.3",
    "moviepy>=2.2.1",
    "pydub>=0.25.1",
    "imageio>=2.37.0",
    "seaborn>=0.13.2",
    "plotly>=6.2.0",
    "streamlit>=1.47.0",
    "streamlit-lottie>=0.0.5",
    "matplotlib>=3.10.3",
    "streamlit-option-menu>=0.4.0",
    "streamlit-card>=1.0.2",
    "opencv-python>=4.11.0.86",
]
